% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{arq_check}
\alias{arq_check}
\title{Check if file was read correctly}
\usage{
arq_check(arq)
}
\arguments{
\item{arq}{name of the file}
}
\description{
Does some tests to check if a file was read correctly.
}
