% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{courts}
\alias{courts}
\title{Court list data}
\format{A data frame with 2941 rows and 4 variables:
\describe{
   \item{id_comarca}{Internal id of the comarca}
   \item{nm_comarca}{Name of the comarca}
   \item{id_court}{Internal id of the court}
   \item{nm_court}{Name of the court}
}}
\source{
\url{http://www.tjsp.jus.br/produtividadeweb/}
}
\usage{
courts
}
\description{
Contains court list data obtained from TJSP productivity site.
}
\keyword{datasets}
