% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{prod_download}
\alias{prod_download}
\title{Downloads productivity pdf files}
\usage{
prod_download(.data, path = "data-raw/pdf", overwrite = FALSE)
}
\arguments{
\item{.data}{database of id's returned by \code{prod_list}}

\item{path}{folder where the files will be saved.}

\item{overwrite}{overwrite?}
}
\value{
tibble informing if each download was OK.
}
\description{
Downloads TJSP productivity pdf files based on database returned by \code{prod_list}.
}
