% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{prod_download_async}
\alias{prod_download_async}
\title{Downloads productivity pdf files, async}
\usage{
prod_download_async(.data, max_nproc = 10, sleep_secs = 1,
  path = "data-raw/pdf", overwrite = FALSE)
}
\arguments{
\item{.data}{database of id's returned by \code{prod_list}}

\item{max_nproc}{maximum number of processes. Default 10 processes.}

\item{sleep_secs}{seconds to sleep between process checks. Default 1 second.}

\item{path}{folder where the files will be saved.}

\item{overwrite}{overwrite?}
}
\value{
nothing
}
\description{
Downloads TJSP productivity pdf files based on database returned by \code{prod_list}.
}
