% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courts.R
\name{prod_download_courts}
\alias{prod_download_courts}
\title{Download court list}
\usage{
prod_download_courts(minimal = TRUE, save = FALSE,
  path = "data-raw/courts")
}
\arguments{
\item{minimal}{download list of courts based in 2016-01 or consider all
possible year-month combinations?}

\item{save}{save intermediate files? Recommended if \code{minimal} is
\code{FALSE}.}

\item{path}{folder where intermediate files will be saved.}
}
\value{
tibble with four columns and variable number of rows depending on
  \code{minimal} parameter.
}
\description{
Download court list from TJSP website.
}
