% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{prod_list}
\alias{prod_list}
\title{Build pdf list}
\usage{
prod_list()
}
\value{
tibble with 6 columns (year, month, id_comarca, id_court, nm_comarca,
  nm_court)
}
\description{
Builds all valid combinations of pdf available.
}
