% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{prod_parse_uni}
\alias{prod_parse_uni}
\title{Court productivity data}
\usage{
prod_parse_uni(arqs)
}
\arguments{
\item{arqs}{character vector contaning the paths of the pdf files.}
}
\description{
Parses pdf file to get Court productivity data. Ignores information about the
judge.
}
